Added multiple features to the control.lua file
The logs will now also output if a user has been banned, unbanned etc. so Jammy can return the result to the Discord chat. Added functions to parse the serverData.json file and turn it into an object, from object to JSON file too. Still working on the production statistics for grafana